add ECDSA tests
