+++ packages/ethereum/test/HoprChannels/types.ts
@@ -1,11 +1,3 @@
-export type Account = {
-  privKey: string
-  pubKey: string
-  pubKeyFirstHalf: BN
-  pubKeySecondHalf: BN
-  address: string
-}
-
 export type Ticket = {
   recipient: string
   proofOfRelaySecret: string
