+++ packages/ethereum/test/types.ts
@@ -1,0 +1,7 @@
+export type Account = {
+  privKey: string
+  pubKey: string
+  pubKeyFirstHalf: BN
+  pubKeySecondHalf: BN
+  address: string
+}
